pyproject.toml
[project]
name = "trmnl-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.5"
dependencies = [
    "pillow>=12.0.0",
]
